# Alternate fix for #261