This should address the perf regression of #263 while allowing visibility through thin leaves as described in #261 (though I don't have the setup to confirm this.)