@rrika 2023-08-11
I confirmed this by doing a "-fast" run of gmsp3v2.bsp

ee256f339e9ec2090757f70f0b3b052c24f32e00: average leafs visible: 1197
74e1a0f6f02a2a469ebf6d7c05c175096cfe01a3: average leafs visible: 1324
backported PR: average leafs visible: 1197

However, this number is provided by LeafFlow which at the time had a bug in counting, and doesn't match newer versions which use ClusterFlow.

```c++
    numvis = 0;
    for (i = 0; i < portalleafs; i++)
        if (outbuffer[i >> 3] & (1 << (i & 3))) // <---- this should be (i & 7)
            numvis++;

    /*
     * compress the bit string
     */
    if (verbose > 1)
        logprint("leaf %4i : %4i visible\n", leafnum, numvis);
    totalvis += numvis;
```

With that fixed it becomes:
ee256f339e9ec2090757f70f0b3b052c24f32e00: average leafs visible: 1217
74e1a0f6f02a2a469ebf6d7c05c175096cfe01a3: average leafs visible: 1346
backported PR: average leafs visible: 1217

And back in the present:
e099c292b05be39f42e64a27336e322eaf2218cb: average leafs visible: 1346
39f3ee7db7b60f168b389a6f93145c4af43b10e8: average leafs visible: 1217

I also found that these days fast vis takes 9.5sec when it used to take 2.5sec. I haven't bisected the reason for that yet.