@rrika 2023-10-20
@Paril since the refactor commits in the range I mentioned at the end are yours, do you have any insight you could share? (I can't build them with gcc/clang, so pinpointing where the slowdown happens would already help)