@Paril 2023-10-22
Hey. I've been a bit busy on NightDive stuff as well as sick this weekend so I haven't had much time to do anything. I'll try to take a look at this this week for you.