@ericwa 2023-10-22
I can take a look as well.. I do get a slowdown for a `-fast` vis of `gmsp3v2.bsp` between those commits but it's not as bad

- 0.9s https://github.com/ericwa/ericw-tools/commit/997a039b65ff04546b6a17a97fff6652004143a8 (good)
- 1.36s https://github.com/ericwa/ericw-tools/commit/c54fea1e8742089c66e788df027af86debe4fead (bad).

I commented out the `Leaf portals saw into leaf` warning in both cases because the spam was influencing the timing.
It seems repeatable.