@ericwa 2023-10-22
> I also found that these days fast vis takes 9.5sec when it used to take 2.5sec. 
> ~I haven't bisected the reason for that yet.~ The perf regression happens somewhere between [7f8d2cf](https://github.com/ericwa/ericw-tools/commit/7f8d2cfb49d2cfee2d752b4cc47b0d4dd0852f27) (good) [997a039](https://github.com/ericwa/ericw-tools/commit/997a039b65ff04546b6a17a97fff6652004143a8) (good) and [c54fea1](https://github.com/ericwa/ericw-tools/commit/c54fea1e8742089c66e788df027af86debe4fead) (bad). I can't build the commits between, there are too many errors.

@rrika do you remember if those 9.5sec vs 2.5sec measurements were debug or release? I'm guessing debug?

I'm still working on bisecting that performance regression, it's kind of a pain because the code was undergoing heavy rewrites in this period and some commits don't work.

- e8bd554cc0666ee21f2fd1c3f4b34091b5d8f9b3 good (0.9s)
- 41b9a79e8388b91749edeee1a4a0ba198cbd121b (Merge branch 'qbsp-contentflags' into type-cleanup) good-ish (1.0s)
  needed patch:

    ```
    diff --git a/include/common/cmdlib.hh b/include/common/cmdlib.hh
    index 92ae6c51..fe4e4dcc 100644
    --- a/include/common/cmdlib.hh
    +++ b/include/common/cmdlib.hh
    @@ -562,7 +562,7 @@ template<typename Dst, typename Src>
    constexpr Dst numeric_cast(const Src &value)
    {
        if (numeric_cast_will_overflow<Dst, Src>(value)) {
    -        throw std::overflow_error("value");
    +        //throw std::overflow_error("value");
        }

        return static_cast<Dst>(value);
    diff --git a/vis/soundpvs.cc b/vis/soundpvs.cc
    index 6c98d5c2..c7e95169 100644
    --- a/vis/soundpvs.cc
    +++ b/vis/soundpvs.cc
    @@ -47,7 +47,14 @@ static aabb3d SurfaceBBox(const mbsp_t *bsp, const mface_t *surf)
            else
                vertnum = bsp->dedges[-edgenum][1];

    -        bounds += bsp->dvertexes[vertnum];
    +        if (vertnum >= 0 && vertnum < bsp->numvertexes) {
    +            const auto &bsp_vert = bsp->dvertexes[vertnum];
    +
    +            auto val = qvec3f(bsp_vert[0], bsp_vert[1], bsp_vert[2]);
    +            bounds += val;
    +        } else {
    +            LogPrint("invalid vertex number! {}\n", vertnum);
    +        }
        }

        return bounds;
    ```
    
    - 9d828d2085ac9b088ebfef00c7fd38cd87d2f3d1 (Structured texture data!) bad (1.3s)