@rrika 2023-10-22
I checked and I was measuring debug builds, on release I get the exact same times (0.9 vs 1.36)