@ericwa 2023-10-22
Awesome, thanks.

I did one more test:

- 997a039b , release build, gmsp3v2.bsp full vis: 126.4s
- c54fea1 , release build, gmsp3v2.bsp full vis: 154.2s
- e84ba4cc68bf64c08ef61ebcfc9f1e1428969fc7 ( current `brushbsp` branch), release build, gmsp3v2.bsp full vis: 158.0s

and repeating the results from above,

- https://github.com/ericwa/ericw-tools/commit/997a039b65ff04546b6a17a97fff6652004143a8 gmsp3v2.bsp fast vis: 0.9s
- https://github.com/ericwa/ericw-tools/commit/c54fea1e8742089c66e788df027af86debe4fead gmsp3v2.bsp fast vis: 1.36s

So between 997a039b and c54fea1 , full vis got 22% slower, fast vis got 50% slower.

(test system is Win 11 64-bit, 3950x, 32 threads).

I was wondering if it was just the bsp loading/saving had gotten ~450ms slower, which wouldn't be so bad, but since the full vis took about 30 seconds longer it points to a perf regression in the vis/winding/clipping code.  I guess this confirms there was a second vis slowdown in Oct 2021 separate from https://github.com/ericwa/ericw-tools/issues/263 and probably related to the C++ifying / refactoring from that time.