@ericwa 2023-10-30
I bisected the ~120s -> ~150s increase for gmsp3v2.bsp vis to a6f6cc2d150493b899886040e170527195e81ab5. It needed to be patched to actually load a q1bsp, here is the patch I used that gets just enough working to load the .bsp and record the vis runtime:
[a6f6cc2d-fixup.patch](https://github.com/ericwa/ericw-tools/files/13200306/a6f6cc2d-fixup.patch)

For reference, here is the fixup patch for the previous commit, 87d31e6d4f5f3b75454f96c31a59e50c3b1b2256, this one took around 120s:
[87d31e6d-fixup.patch](https://github.com/ericwa/ericw-tools/files/13200325/87d31e6d-fixup.patch)


Next step is to dig into the a6f6cc2d changes to see exactly what caused the slowdown.