@rrika 2023-11-01
I'll take the a6f6cc2 changes apart a bit.

I did however notice a 1 sec slow-down on fast debug vis even before 87d31e6.

I experimented with creating something between the old winding_t and the new polylib winding by doing:
```diff
--- a/include/vis/vis.hh
+++ b/include/vis/vis.hh
@@ -38,7 +38,7 @@ struct winding_t
     int numpoints;
     vec3_t origin; // Bounding sphere for fast clipping tests
     vec_t radius; // Not updated, so won't shrink when clipping
-    vec3_t points[MAX_WINDING_FIXED]; // variable sized
+    qvec3d points[MAX_WINDING_FIXED]; // variable sized
 };
 
 winding_t *NewWinding(int points);
diff --git a/vis/vis.cc b/vis/vis.cc
index a7f71d21..62e0665f 100644
--- a/vis/vis.cc
+++ b/vis/vis.cc
@@ -302,7 +302,7 @@ winding_t *ClipStackWinding(winding_t *in, pstack_t *stack, plane_t *split)
     neww->radius = in->radius;
 
     for (i = 0; i < in->numpoints; i++) {
-        p1 = in->points[i];
+        p1 = &in->points[i][0];
 
         if (sides[i] == SIDE_ON) {
             if (neww->numpoints == MAX_WINDING_FIXED)
@@ -323,7 +323,7 @@ winding_t *ClipStackWinding(winding_t *in, pstack_t *stack, plane_t *split)
             continue;
 
         /* generate a split point */
-        p2 = in->points[(i + 1) % in->numpoints];
+        p2 = &in->points[(i + 1) % in->numpoints][0];
         fraction = dists[i] / (dists[i] - dists[i + 1]);
         for (j = 0; j < 3; j++) {
             /* avoid round off error when possible */
```
This slows things down by a whole second. Does that reproduce for you?