@ericwa 2023-11-09
I think I'm done fixing the vis perf regressions on the `brushbsp` branch, aside from the one this PR itself addresses. (Debug builds are still going to be slower than before the `vec3_t` -> `qvec` conversion.)

To recap the recent commits that made a difference: (same benchmark as before, 32-thread vis on `gmsp3v2.bsp`, RelWithDebInfo build on Windows / MSVC)
- e9532346c7dd29bab088f5ce41b52203d160ad92 153s -> 141s
- c9edf4ddaa46cc3bdc62046e643c59fdca7c3f92 157s -> 61s
- 54e7aadda7b57100bb94f12d57abbca79520a13d 59s -> 56s
- 8845866c00fd15e9b91efe0fe6a705ffee098169 56s -> 52s

The last two might be MSVC specific. The biggest win was getting rid of the global stat counters (c_chains, etc.) and moving them to per-thread.

So finally it should be a good time to look at this PR, sorry it took so long. I'd like to dig into BasePortalThread and understand what's going on better - when I patched https://github.com/ericwa/ericw-tools/issues/261 I didn't really understand why the "fix" worked.